refactor(web): migrate jobs into src/domains
Third domain migrated to the domain-driven structure. Co-locate the job-board
feature under apps/web/src/domains/jobs:
- components/ <- components/jobs/* (JobList, JobForm, JobDetail, JobsFilters)
- server/ <- lib/jobs/* (create/update/delete-job, source-intake)
- index.ts barrel + README; repoint all 8 importers at @/domains/jobs

Validation schemas remain in lib/schemas/jobs (shared with API routes and AI
tools), consistent with the other migrated domains. Typecheck, lint, and build
pass.
