+++ apps/web/src/app/[orgSlug]/jobs/page.tsx
@@ -4,8 +4,7 @@ import { getOrgContext } from "@/lib/auth/roles";
 import { createClient } from "@/lib/supabase/server";
 import { getTranslations } from "next-intl/server";
 import { PageHeader } from "@/components/layout";
-import { JobList } from "@/components/jobs/JobList";
-import { JobsFilters } from "@/components/jobs/JobsFilters";
+import { JobList, JobsFilters } from "@/domains/jobs";
 import Link from "next/link";
 import { Button, EmptyState } from "@/components/ui";
 import { sanitizeIlikeInput } from "@/lib/security/validation";
