+++ apps/web/src/domains/jobs/components/JobDetail.tsx
