+++ apps/web/src/domains/jobs/server/delete-job.ts
